completed navbar
