+++ index.html
@@ -1,11 +1,11 @@
 <!doctype html>
-<html lang="en" class="h-full">
+<html lang="en" >
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Freelance Time Tracker</title>
   </head>
-  <body class="h-full flex-col">
+  <body>
     <div id="root"></div>
     <script type="module" src="/src/main.jsx"></script>
   </body>
