--- src/App.css
@@ -1,39 +1,0 @@
-@import "tailwindcss";
-@import "tailwindcss/preflight";
-@tailwind utilities;
-
-*,
-*::before,
-*::after{
-  box-sizing: border-box;
-}
-
-:root {
-  font-family: system-ui, Avenir, Helvetica, Arial, sans-serif;
-  line-height: 1.5;
-  font-weight: 400;
-
-  color-scheme: light dark;
-  color: rgba(255, 255, 255, 0.87);
-  background-color: #242424;
-
-  font-synthesis: none;
-  text-rendering: optimizeLegibility;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-}
-
-header{
-  position: sticky;
-  z-index: 1000;
-  top: 0rem;
-
-}
-
-body {
-  margin: 0;
-  display: flex;
-  place-items: center;
-  min-width: 320px;
-  min-height: 100vh;
-}
