--- src/components/Logo.jsx
@@ -1,10 +1,0 @@
-import logo from "../assets/img/logos/TimeTrackerLogo-bgDark.jpg";
-
-
-export default function Logo() {
-  return (
-    <div className="">
-      
-    </div>
-  );
-}
