+++ src/components/navbar/navbar.css
@@ -1,0 +1,61 @@
+
+
+.nav {
+  padding: 1rem 3rem;
+  margin-inline: 0;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  width: 95%;
+  height: 5%;
+  
+  background-color:rgba(48, 49, 50, 0.912);
+  border-radius: 0 10, 10, 0;
+
+  position: fixed;
+  z-index: 1000;
+  top: 0;
+  left: 0;
+
+
+  ul {
+    display: flex;
+    gap: 2rem;
+    list-style: none;
+    margin: 0;
+    padding: 0;
+
+  }
+
+  li {
+    color: inherit;
+    text-decoration: none;
+    font-weight: 500;
+  }
+}
+
+.navbar {
+  display:flex;
+  justify-content: space-between;
+  align-items: center;
+  width: 100%;
+  font-size: 16px;
+  font-weight: 600;
+
+  color: #ffa500;
+}
+
+.nav-link {
+  color: rgb(12, 14, 133);
+}
+.nav-link:hover{
+  color: #ffa500;
+}
+
+.nav-img {
+  display: flex;
+  height: 50px;
+  border-radius: 100%;
+}
+
+/* <--------- End Navbar ----------> */
