+++ src/main.jsx
@@ -2,9 +2,8 @@ import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import { Provider } from 'react-redux';
 import { store } from './app/store.js';
-import './index.css'
 import App from './App.jsx'
-
+//import './index.css'
 createRoot(document.getElementById('root')).render(
   <StrictMode>
     <Provider store={store}>
